Share single Spring context across the beans integration tests
